[tool] Fix issue with spaces between suppressions
Summary:
When adding multiple suppressions to a JSX container we would add an empty line between the suppression which means the suppression at the top would not work. e.g.

```
<div>
  {thing}
</div>
```

Turns into:

```
<div>
  {
     // $FlowFixMe[code1]

     // $FlowFixMe[code2]
     thing}
</div>
```
The logic currently assumes the JSXExpression is on a single line, the fix was to check if the opening expression element is on its own line and not add a newline after if thats the case.

Reviewed By: evanyeung

Differential Revision: D38294092

fbshipit-source-id: cfae736606b40cdb6ea975c9f8b73922b90eb11e