[tests] Stop testing constrain writes env
Summary:
Changelog: [internal]

This is a part of a stacked diff to drop support for classic-env, constrain writes and new-env, which I plan to land after the version cut of the last version we intend to support these old envs.

In this diff, I removed the flag to test constrain writes, and constrain writes specific overrides. This means that we will need to rename .cw override to .new_env override, and from TODO_constrain_writes to TODO_new_env.

I also deleted the constrain_writes_dir suite, since that will be a mode no longer supported.

Reviewed By: panagosg7

Differential Revision: D38301249

fbshipit-source-id: 852d88c8cf9ef8262b6038894140abc0fd583762