fix file descriptor leak on Windows
Summary:
on Windows, workers are spawned for each job. since each job is a fresh connection, those connections need to be closed.

on Mac/Linux, they are prespawned at startup and forked for each job. the fork dies but the worker itself lives forever and so doesn't need to clean up.

when jobs are cancelled (e.g. cancelled recheck), the send/read Lwt promises in WorkerController can raise `Lwt.Canceled`. this currently causes us to skip closing the worker's file descriptors. eventually, enough cancelled rechecks causes Windows to run out of file descriptors and die with a "Too many files open" Sys_error when spawning a worker.

so we need to call `close` when cancelling. to be safe, I decided to use the `noerr` version so closing doesn't itself raise a `Sys_error`.

Changelog: [fix] Fix a bug on Windows that could lead to a crash if files are changed faster than they can be rechecked

Reviewed By: samwgoldman

Differential Revision: D38365630

fbshipit-source-id: 47b2d351d66652ba8a6525d88d9a0afd1af84698