[tests] Re-record or delete some test suites marked with TODO_new_env
Summary:
Changelog: [internal]

This is a part of a stacked diff to drop support for classic-env, constrain writes and new-env, which I plan to land after the version cut of the last version we intend to support these old envs.

In this diff, as a prep step to stop testing new-env, I audited all the tests marked with TODO_new_env.

- `this` suite is already passing in new-env, but has some minor error message difference in resolved_env, so I added a resolved_env override.
- `closure` has some minor message differences between classic and new-env on reading and writing to unknown globals, so I added a new_env override.
- `extends` and `more_path` are two known cases where we decide not to add post inference check for instanceof tests, so I added their new_env override.

Reviewed By: panagosg7

Differential Revision: D38301310

fbshipit-source-id: 32add9deb960ff74ec3cfe568f5d9b0a137f5571