[tests] Stop testing new-env
Summary:
Changelog: [internal]

This is a part of a stacked diff to drop support for classic-env, constrain writes and new-env, which I plan to land after the version cut of the last version we intend to support these old envs.

In this diff, I removed testing on new-env, and mostly renamed .new_env override to .resolved_env override. Some of them have minor test message differences.

Reviewed By: panagosg7

Differential Revision: D38301371

fbshipit-source-id: 3bebbec37f47e916462a515ba3777378bb3cfe85