Rename env_mode flag to inference_mode
Summary:
This renames the `env_mode` flag to `inference_mode` and restricts the publicly available modes to:
1. classic (old-env)
2. constrain_writes (Reordered env)
2. experimental.lti (enforced env)

Changelog: We are releasing a new inference mode that marks significant progress on our way to local type inference. You can enable this new mode in your codebase by adding `inference_mode=constrain_writes` to the `[options]` section in the flowconfig. Look out for a blog post explaining the new behavior and the codemods you can use to get rid of errors associated with it.

Reviewed By: SamChou19815

Differential Revision: D38332142

fbshipit-source-id: 57fdd13afe7a870e4cb2ae77c4030b0f32397ac4