remove unused LspProt EOF message
Summary:
when the monitor is exiting, it sends a `NotificationFromServer (ServerExit ...)` followed by a `NotificationFromServer EOF` to the lsp process.

since they are always sent together, we don't need the `EOF`. flowLsp already handles the `ServerExit`, so raising an exception on the `EOF` just causes log spew and potentially Lwt races.

Reviewed By: SamChou19815

Differential Revision: D36548056

fbshipit-source-id: 8aa9fd88016e19dffcb345fbe82d608e4cae8a45