Flow any to tvars with import and get_prop error
Summary:
Changelog: [errors] Additional `value-as-type` errors might be emitted, as a result of existing importing errors.

Enforced-env will require that all tvars are resolvable. It will be hard to distinguish whether the tvars are unresolved for valid reasons: e.g. when there are errors. However, in these cases, we should just flow `any` to those touts, so that `TvarResolver` doesn't need to special case when error occurs.

Reviewed By: panagosg7

Differential Revision: D38394827

fbshipit-source-id: f4cab1551f20e6f62763b91acb18bac32b210d17