fix lsp test runner when restarting server
Summary:
the LSP test runner captures LSP messages per "step". at the beginning of each step, it clears all of the previous messages.

but if it has to start a server, it does that first. that causes messages related to starting the server or received while starting the server to be cleared instead of being part of the step that starts the server.

Reviewed By: SamChou19815

Differential Revision: D38339771

fbshipit-source-id: e8f703a3355578e7411bdfbf84784c8aa00e0cec