[resolved-env][debug][EZ] Add more helpful missing roots information to crash message
Summary:
Changelog: [internal]

I'm debugging some missing dependency issue again, and I found it's very frustrating to scan through a giant list of roots and all, just to find some roots that are missing. The list of missing nodes can be easily computed, so let's just compute it and include it in the error message.

Reviewed By: panagosg7

Differential Revision: D38413580

fbshipit-source-id: 30406648d2810c77eeae67f831ed4ec0aaae1bfd