[name_resolver] Eliminate a write entry created for sentinel refinement that we never write to and read from
Summary:
Changelog: [internal]

First some background on what's happening before this diff.

For sentinel refinement like `obj.type === 'Foo'`. We will create a sentinel refinement. For this sentinel refinement, we will also create a post inference check to ensure that `Foo` is indeed a subtype of `obj.type`.
When we first implement the post-inference check, we are ambitious and decide to support the post inference check for all kinds of `obj`, instead of a simple identifier limited by old-env. For this, we will need to record the type of `obj` as a write, and in statement.ml call `Env.record_expression_type_if_needed`.
However, this causes all sorts of additional errors that's hard to eliminate, so we decide to simply replicate the simple old-env behavior: only add post-inference check for `simple_identifier.type === 'Foo'`. We removed the `Env.record_expression_type_if_needed` call in statement.ml, but we never removed the write entry.

The write entry is mostly harmless so far. It's never written to and never read from. However, as I'm adding missing dependency edges for refinements, I noticed this problem and fixed it in this diff.

Reviewed By: jbrown215

Differential Revision: D38429476

fbshipit-source-id: 23843ba587e941c0e2f0a8b51a4e31cec9936d83