[enforced-env] Make enforced env turn on multiple behaviors designed for LTI
Summary:
Changelog: [internal]

We built exact empty objects and additional LTI annotation enforcements, currently under flags, for LTI. So far, these behaviors are orthogonal and don't interact with each other.

However, as we enter the final phase of LTI implementation with enforced-env, it doesn't make sense to delay the inevitable: these behaviors must be on when enforced-env is released. We can debate when we should flip the default on those flags and remove them, but we can easily make enforced-env automatically enable these behavior for now, since enforced-env is not released.

Reviewed By: jbrown215

Differential Revision: D38374904

fbshipit-source-id: cd568c7f5824b10dbe6ce72a2a0d009fafd6a411