[name_resolver] Do not create a write env entry for delete
Summary:
Changelog: [internal]

For enforced env to work, we need to enforce the invariant that for all write entries generated by name_resolver, there will be an element in name-def that resolves it.

This is currently not the case for the `delete foo.bar` write entry. It's not resolved by any name-def element, but it's also not necessary. Nothing will read from the entry, because name_resolver already decides that a read of `foo.bar` followed by `delete foo.bar` will be `Val.undefined`, which is not associated with any env entry.

Therefore, we should avoid creating this useless write env entry.

Reviewed By: jbrown215

Differential Revision: D38443984

fbshipit-source-id: 9a1eb175c5ed1efa6fa481e17bf1b372cabfa4b7