[providers] fix env when computing empty array providers
Summary:
Before this change, when looking up the env for an entry `x` on `x.push(..)` in the code
```
function foo() {
  var x = [];
  x.push(42);
}
```
we would only look up the toplevel env and miss the definition of `x`. This would lead
to eventually missing `x.push(42)` as a provider for `x`.

This fixes this by looking up the most recent env (`acc`), instead of the one provided
when creating the instance of `find_providers`.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D38341708

fbshipit-source-id: d640f3cb83f80d95e228d078531b980a512c2d03