Fix newline bug in Flow comment suppresion tool
Summary: This diff fixes a bug in the Flow comment suppression tool where it would insert a newline before all Flow errors, even if the error wasn't suppressed. This occurred when a specific error code was given to suppress but only different Flow errors existed in the code. By not touching the AST when there are no error codes, no newline is inserted.

Reviewed By: mroch

Differential Revision: D38473006

fbshipit-source-id: 5cffcaf59f550cdb62379c4809634fa95dea78da