recheck status >> command status
Summary:
the `Finishing_up` event is used both for finishing a recheck and for finishing a request. this diff creates a `Handling_request_end` instead, so that we can ignore status updates about requests when the server is rechecking. requests that happen in parallel should not override the status of the recheck.

(if we did want to show request statuses during a recheck, they would need to be tracked separately from the recheck status. right now, there's just one server status, so when a request finishes, the status bar says the server is done even when the recheck is still ongoing. it would need to be a stack or something instead)

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38485463

fbshipit-source-id: dc5eb2278c1407e27749deaf5b29e04db269bbf9