explicit telemetry lsp event
Summary:
the server sends data about rechecks and other commands to the LSP process, which the LSP process uses to log what else the server was doing that delayed each LSP request.

right now, it does that using ServerStatus, which updates the status bar. it has to be a "significant transition" which means we sometimes get super fast (like 2ms) flickers. and since it runs on parallelizable commands, it interrupts the status of a parallel recheck and makes it say "finishing up" when it isn't.

instead, create a dedicated LspProt message to tell the LSP process about these summaries. in a followup, I'll make `Finishing_up` not a significant transition.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38485462

fbshipit-source-id: 91565bcb903f5dee55508d03124efae52d7cdbf1