starting/finishing typecheck events are not significant
Summary:
now that `Finishing_typecheck` is not used to ferry profiling data to the LSP process, it doesn't *have* to be a significant event (significant events are pushed to the client immediately, instead of throttling to every 500ms and skipping things that take less than that).

this avoids status bar flickers, because this event is usually <10ms.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38485460

fbshipit-source-id: aea0fd9ad2c4d11f9aa9ccc000a59048012dbf67