[tests] Do not enable const params in env_builder_refinement_tests
Summary:
Changelog: [internal]

I'm debugging an issue and found which write to read in new-env disagrees with the result in these tests. Then I found out that const params option is enabled, which explains the behavior difference. There is no value with const params on in these tests, so I flipped it to avoid future confusion.

Reviewed By: jbrown215

Differential Revision: D38481104

fbshipit-source-id: 4e7606d18db2e3dd08e330c6ad8e5264ed460850