No longer use unsealed objects for computed properties in object literals
Summary:
Previously when creating objects with computed properties:
```
const o = {
  [foo]: 1,
};
```
We first created an unsealed object, then wrote the element (e.g. `obj[foo] = 1`), sealed it, and then spread it into the object.

Instead of creating an unsealed object, writing to it, and then sealing it, I create an object directly. This is then spread as before.

Two diffs of interest (by jbrown215):
- D16968206 (https://github.com/facebook/flow/commit/d5eee1554693ee50e909ca5c46056dbb393bacae): "Add indirection to computed properties"
- D16970914 (https://github.com/facebook/flow/commit/40ab881369beb7510664d63323ea5b7e4b245ace): "Restrict computed properties to single lower bounds"

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D38377875

fbshipit-source-id: e6276acfb1b7433c81187e52eb469e9578c94691