[env_builder][EZ] Give global exports its own def_loc_type
Summary:
Changelog: [internal]

Right now we give the global exports write a synthetic loc at `(filename, none loc)`. This can be very tricky everytime we reason about whether there are other constructs that use the same location, so let's just give it its own kind to remove the confusion once and for all.

I name the kind `GlobalExportsLoc` instead of just `ExportsLoc`, because we have `internal_name "exports"` (which sadly still depends on old-env) that happens both in a module and in scoped declare modules.

Reviewed By: panagosg7

Differential Revision: D38542807

fbshipit-source-id: c0c8f8f0ff204a7ecc006bf402f3a2d4c22f9b06