Constrained write error messages for constrained arrays
Summary: Now that we've observed that constrained arrays is a bigger deal than we thought, we should improve error messages to support them. We reuse the `ConstrainedAssingment` use op frame, which we normally use for constrained writes to variables and which points to providers and suggests annotations. We also need to make sure that, given the use_op modifications that we perform in the post-inference path in order to select useful ops, we propagate this frame onto the actual use op that we ultimately choose, which involves a decomposition and inspection of use ops.

Reviewed By: panagosg7

Differential Revision: D38404287

fbshipit-source-id: 5a0b5572ed73e975981e511c4a0f1f7da2185ca0