use existing ast in get-def
Summary:
get-def uses Scope_builder and Scope_builder requires an AST, not a Typed AST. previously we were converting the TAST back to an AST, but the TAST is generated from the AST in the first place! just pass it along. we do this in lots of places already.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38538297

fbshipit-source-id: f09accdf896f62ad756eaeae502d4f4413e7feeb