Use Pierce in flow_js.ml when in Enforced env mode
Summary:
The only precondition to running Pierce in our system is that the args and function must be fully resolved. That is the case in the enforced environment, so we should start using pierce there.

We achieve this via the ImplicitInstantiationKit functor, which uses instantiate_poly when the env is not enforced and Pierce otherwise.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38214221

fbshipit-source-id: 3ddd99a2997584c8c8988e4787f0fcf70786ad35