propagate hints in conditional expressions
Summary:
In expressions of the form `cond ? e1 : e2` we currently did not
propagate a contextual hint *h* to any of the subexpressions `e1`
and `e2`. This diff changes this by propagating
* *h* to `e1`
* *merge(h, `typeof e1`)* to `e2`

The *merge* operation, which is also in this diff, selects its first argument
if it is a meaningful hint (for now this is any case other than `Hint_None`)
and its second argument otherwise.

This diff applies this change in both statement.ml and name_def.ml.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D37041562

fbshipit-source-id: 66d804bfddd939a421f849a8924e8dbf744b74b6