[env_builder] Error on rebinding or reassigning imports
Summary:
Changelog: [internal]

When I am trying to make the resolved-env the default, I found that we are not erroring on reassigning or rebinding imports... This diff fixes that.

We need to visit the import binding identifier with `kind:Ast.Statement.VariableDeclaration.Const`, so it's not considered as an assignment by name_resolver. In addition, we need to treat `Binding.Import` the same as `Binding.Const` when erroring on rebinding or reassignment.

Reviewed By: panagosg7

Differential Revision: D38599204

fbshipit-source-id: b081f8e3c97a0326724af49d7b02c88d78e579eb