[enforced-env] Relax enforcement during InitLib phase
Summary:
Changelog: [internal]

This diff gets rid of most libdef enforcements errors and greatly improves our DX when working on enforced-env.

We need to relax the sealed tvar enforcement for globals during InitLib phase, because libdefs are checking in an arbitrary order with no respect of dependency.
For example, we will check the builtin libdefs in alphabetical order, like `bom.js, core.js, cssom.js, ...`, but it's obvious that `bom.js` depends on `core.js`.
To accomodate this behavior, we already relax the builtin lookup rule, so we shouldn't apply the strict sealed tvar enforcement rule to libdef init phase.

Reviewed By: mvitousek

Differential Revision: D38580107

fbshipit-source-id: 78a8384c3fc8de1b786577da12dad9aa9707dd9c