simplify CLI arg errors
Summary:
when the CLI arguments are invalid (e.g. unknown flag, an invalid enum value, etc), just show the error message and not the dozens of lines of the full `--help` output. the help output buries the error, and is especially useless in `--json`.

while I was here I also improved the error messages so they read better

as a side motivation, the `--help` output includes the executable name, which was being included in the `json_exit` test. it expected `flow`, but it's `flow.exe` on Windows and/or when building with dune, making this test fail. instead of fixing that, just stop doing it.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38665867

fbshipit-source-id: 1dceab35f43c2c2a217e21c19f2013fbbc12c17b