[enforced-env] Declare dependency on refinement writes when visiting expressions
Summary:
Changelog: [internal]

When name_def sees an declaration like `const a = expr`; It will find out what writes `a` depend on by visiting `expr`. If it finds a `pattern_identifier` inside expr (e.g. `const a = b = 3;`), it will add the dependency on that `pattern_identifier` (in this example `b`).

The same logic has not been implemented for heap refinement writes. Consider

```
const a = foo.bar != null ? foo.bar : "default";
```

Similar to simple identifiers, the dependency analysis should figure out that `a` depends on the write `foo.bar` (from `foo.bar != null`).

This diff implements the logic by trying to find such writes when we visit expressions. I factored out the logic to check for the existence of a legal write at `(kind, loc)` to a helper function, and then use it to find refinement writes.
We also need to be careful not to introduce self-cycles.
e.g. To resolve a `MemberAssign` in
```
foo.bar = 3;
```
we need to visit the LHS expression to find some potential dependency on foo write, but we should not be recording a dependency on `foo.bar.`

Reviewed By: mvitousek

Differential Revision: D38430309

fbshipit-source-id: 54b427a83d187bcde4c9113f441617101ddcd2db