[reordered-env] Use argument location instead of the entire pattern location for function parameter names
Summary:
Changelog: [internal]

This diff keeps the location that we produce missing-local-annot errors in sync with the old environment: on the argument rather than the entire pattern.

e.g.

- In `...rest`, error on `rest` rather than `...rest`
- In `a=3`, error on `a` rather than `a=3`

This bug is preventing us from switching to reordered-env for annotate-lti codemod due to error location differences. Since the bug is fixed, I switched annotate-lti codemod to use reordered-env.

Reviewed By: gkz

Differential Revision: D38712297

fbshipit-source-id: 1014d28c4b8e27527f349598b5ad27d4f340533f