cache empty array literal type generated in New_env.init_entry
Summary:
`New_env.init_entry` is somewhat unique in the way that it handles empty array literal
since it does the checking of the literal itself. This is typically done in env_resolution.ml
or statement.ml, and the results are subsequently cached in the node cache. Before,
this caching wasn't happening, so when we visited the array from statement.ml
we would raise an `empty-array-no-provide` error again. This diff fixes this by caching
the type of the array in `New_env.init_entry`. To do this, we add a location field in
`EmptyArrayWrite` so that we can reconstruct the empty array AST node.

Reviewed By: SamChou19815

Differential Revision: D38668261

fbshipit-source-id: c459d3e6820bb9d957212ebd8b9162d294253c7a