[resolver] improve hint in resolve_binding_partial Root Value
Summary:
By passing `dumy_hint` to each expression under `Root (Value _)`, we
effectively assumed that each "root" expression was passed a type hint.
This is not accurate. This diff fixes this by determining whether the expression is
a provider or not.

As a general rule of thumb, an expression has a hint iff it is not a provider. This
rule needs to be adjusted a bit for the case of empty array literals, because
of the nuanced way that we determine the type of an empty array literal based on
downstream uses. We consider that a hint for an empty array exists, iff there is a
non-zero number of providers associated with this literal.

This diff still does not compute the type of the hint based on providers. This is still
a TODO, but by passing down `Hint_None` when a hint won't be present, it
unblocks various codemods.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38668260

fbshipit-source-id: dd2cfd78a27275a540adc94669ac073fdd4f4279