(BETA) Rank Autoimport Suggestions by count and fuzzy_score
Summary:
Autoimport Suggestions are sorted with respect to each other by this score: count + fuzzy_score^4 * 1000 + kind of import. The higher the score, the higher their rank should be. VScode does its own sorting and filtering so it is difficult to control the exact order without turning off VScode's sorting/filtering, but for any 2 imports with the same name: we are guaranteed that an import that appears more in the codebase will rank higher than an import that appears less. This diff maintains the current behavior of ordering where non-imports < imports < utilities.

Changelog: [internal] (Beta) Autoimport suggestions now use count in its ranking algorithm, not yet rolled out.

Reviewed By: mroch

Differential Revision: D38021433

fbshipit-source-id: 01e762f188926734b79fad45429a194b19e046c8