Counting Globals
Summary:
We are able to use scope analysis to find all the Globals that a file uses. Using this information we can add the usage count of a Global across the codebase into our export_index. Originally, we were not counting globals so we could not compare them to exports, but now we can.

Changelog: [internal] Globals now have their correct count stored in export_index

Reviewed By: samwgoldman

Differential Revision: D38661746

fbshipit-source-id: 0b73a47f54fd7b61412fed49b3c9a13ce97fc730