Resolve unannotated parameters
Summary:
This diff ensures that unannotated parameters, which are errors, still can be resolved to some type. It does so by ensuring, both when resolving a function sig and when resolving a `Contextual` name_def, that the tvar for a parameter gets `any` as a lower bound if the parameter is unannotated and the hint is `Hint_None`.

In order to do this, we now need to pass a hint through the `Function` name def, in order to distinguish whether a function is contextually typed or not. We can then inspect the hint to decide whether to resolve the `FunT`'s parameter types to `any` or not.

Reviewed By: SamChou19815

Differential Revision: D38685043

fbshipit-source-id: c85328bb9ad27354762bb897554fca1978f84038