Name def support for declared modules
Summary:
This diff builds support for declared modules as their own name def in the resolved environment. This requires some minor fixes to the AST mapper--a declared module depends on all the defs defined within it, which may include imports; we need to be careful about which identifiers from an import statement we're inspecting in determining the ordering. (This was not exercised before because an import statement was never syntactically included within some other name def previously.)

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38692363

fbshipit-source-id: e28e2e44555f64ffa28d3bafa39f9c14fa91f6c8