Resolve unannotated fields
Summary:
Under the resolved environment, we allow fields initialized to literals and to annotated functions to not have annotations on the property name. This diff moves the resolution of said fields to mk_class_sig in order for them to be resolved along with the overall class, rather than being unresolved tvars when we seal the class' types.

This diff also resolves illegal unannotated fields to `any` (but only if the enforced environment is enabled, so as to not modify existing behavior). It also adds an error on such cases: this error currently entirely overlaps with the error generated in type_inference_js for Class name_defs with missing annotations, but in the future it would be preferable to raise such errors at the point where they're checked, rather than as a side effect of the name_def process. Once corresponding errors are added for all cases (which will also need to be resolved to `any` for the enforced environment) I'll disable errors being added in type_inference_js.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38677190

fbshipit-source-id: 29f1a169191e9347430d02a4b415174040facccc