Propagate any to touts for catchall errors
Summary:
This diff makes progress (but doesn't totally finish--see the `_TODO`s in default_resolve.ml, below) on ensuring that touts of incompatible constraints will be resolved to any, rather than being empty tvars and leading to tvar resolution errors in the enforced environment. The cases relevant here are the catchalls of flow_js: when we hit the bottom of the pattern match and emit an `EIncompatible` error. When we do so, we'll now inspect the upper bound for touts, and flow `any` into them as needed. (We also do so when no branch of an intersection matches.)

In addition, we now admit a little bit more any propagation, in order to allow these new `any`s to propagate through constraints to populate tvars. This will also likely need to be expanded.

Finally, this diff now means that a lot of cases will have new `any`s flowing into `TypeT`s: whenever an mismatch happens and we propagate an any, if the result was to be used as an annotation somewhere (e.g., if we tried to pass too many/few tparams in to a PolyT class type annotation), rather than being an empty tvar we'll now create an `any ~> TypeT` flow. We currently treat most such flows as errors, but doing so now creates a lot of extra errors that are always downstream of other errors. To resolve this, we now only generate value-as-type errors from `AnyT`s *without* the `AnyError` source. The result is a slight decrease in the number of value-as-type errors, because there were pre-existing

Reviewed By: SamChou19815

Differential Revision: D38692849

fbshipit-source-id: abf48f27ff94fb9d2a5b80c154ea36c913d6b1aa