Resolve unannotated method returns
Summary:
This diff ensures that when the sealed tvars mode is enabled, any method return types that are unannotated will be resolved to `any`. It also adds missing-local-annotations to these cases, which currently overlap with the errors added in name_def.

This diff also builds better support for constructors. Previously, constructors were never required to be annotated, but in fact a constructor that actually returns a value should be annotated, as it forms part of the class signature. Most constructors don't explicitly return, and they still don't need to be annotated, but this diff builds out full support for erroring on unannotated constructors that explicitly return. Along the way, as a nice(?) bonus, we can now report when constructors are defined illegally (as a generator or async, or theoretically as a predicate, although predicate methods don't parse).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D38687457

fbshipit-source-id: 9bf2626f154b690ae46cd4be0762ef0aa6647e3d