Propagate any through TypeT to populate tvars
Summary:
This diff changes the behavior of the `any ~> TypeT (t)` constraint to propagate `any` into `t`. This was necessary to have some erroneous operations still result in nonempty tvars: for example, a type annotation of an undeclared type would be (in addition to an error, of course) an unresolved empty tvar if the `any ~> TypeT` flow cut off further propagation.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D38693300

fbshipit-source-id: ab48b7bbd5207cac537f6e36e9e987c582ee8819