[enforced-env] Resolve function this type together with function names
Summary:
Changelog: [internal]

Similar to the logic that resolves all kinds of this and super when we resolve a class, we should resolve function this type that is associated with a function when we resolve a function. We can do that by attaching whether a function also defines `this` to the `Function` def, and then asks the tvar resolver to resolve the function this at the function loc.

Not all functions define `this`. Arrow functions don't, and `this` in objects are considered unsupported and illegal.

Reviewed By: mvitousek

Differential Revision: D38673408

fbshipit-source-id: 5f42458b8f63cfac59810e056f3ca4e5068b3f83