Bugfix in counting Named, Namespace, Namedtype and Default imports
Summary:
The counts for Namespace and Namedtype were getting mixed up. I also realized that the Default imports were stored in the list of Named Imports so I am able to count Defaults correctly now. Originally we thought that if an import statement was "empty", it was importing default, but that is not the case.

Changelog: [internal] Named, Namespace, Namedtype, and Default imports are now being counted corrrectly.

Reviewed By: SamChou19815

Differential Revision: D38670771

fbshipit-source-id: 07ec277290626b46fe359ad6a4042100d34f8609