[env-resolver] Add hint to Root (Value _)
Summary:
This diff adds a `hint` field to `Value` `root` of name_def.ml. This is used in the env
resolver to properly the RHS of assignments using the available hint.

This improves on the approach of D38668260 (https://github.com/facebook/flow/commit/87dd2ff08cf96a91bc8c87689b377fe1aa15b3a8) that did not propagate the hint correctly
to the `[]` in:
```
declare var v: Array<number>;
v = [];
```

Reviewed By: SamChou19815

Differential Revision: D38811413

fbshipit-source-id: fc6a16a1acc6e86c62e22dbfdd41abd7355782d5