support flow.suggest.rankAutoimportsByUsage vscode setting
Summary:
adds support for a `flow.suggest.rankAutoimportsByUsage` vscode setting, which can have the values `"default" | "true" | "false" | true | false`. `"default"` or null or unset means that it should obey the flowconfig, `true` forces it on, and `false` forces it off.

this will let people opt in or out of this feature while we're experimenting with it.

Changelog: [internal]

Reviewed By: azhu01

Differential Revision: D38884951

fbshipit-source-id: 7e2d8f44e0fc020ae0f45203db47ae810e7b7fae