[new-env] Bind function_this in `Func_sig.functiontype` and `Func_sig.methodtype`
Summary:
Changelog: [internal]

In this diff, I changed the location we bind function level this type from `Func_sig.toplevels` to `Func_sig.functiontype` and `Func_sig.methodtype`. Conceptually, this is the right thing to do, because `this` type is part of the function signature, while `Func_sig.toplevels` deals with checking the function body.

This diff is mostly a refactor by moving `Env.bind_function_this` around and passing the this location to support the refactor. However, it has a profound impact of dramatically increasing the number of successful tests under enforced-env, because `env_resolution` uses `Func_sig.functiontype` to resolve annotated functions, and this refactor allows us to find the type of function type and function this type together.

Reviewed By: panagosg7

Differential Revision: D38721543

fbshipit-source-id: 484a097d838a078f7c2de739e58b1eb38174be7a