[refactor] Do not return function this type from `Func_sig.toplevels`
Summary:
Changelog: [internal]

Currently, `Func_sig.toplevels` will return the this type of a function. `Func_sig.toplevels` is called in two different places:

1. In class_sig, where the the returned this type is ignored.
2. In statement.ml, where the returned this type is thread into `Func_sig.functiontype`.

Now let's see whether `Func_sig.functiontype` actually needs the this type computed by `Func_sig.toplevels`.

According to the code:

https://github.com/facebook/flow/blob/7e16bed9db3390e53fe4d3faad21737074ef14f0/src/typing/func_sig.ml#L168

All we care about is the default this value, which is `mixed` when the body doesn't contain `this` or a tvar if the body contains `this`. `Func_sig.functiontype` will use the this annotation by itself. Therefore, all we need to pass in is a default this type, instead of the this type processed by `Func_sig.toplevels`.

The above observation allows us to greatly simplify the code:

1. In the previous diff, `Func_sig.toplevels` will no longer bind function this, so we can simplify the `default_this_type` variant to be `Type.t option`, where `Some t` represents the existence of function this, and `None` represents that we should inherit this from the upper scope.
2. We will pass a default this type directly to both `Func_sig.toplevels` and `Func_sig.functiontype`, so that `Func_sig.toplevels` no longer needs to return a processed this type.

Reviewed By: mvitousek

Differential Revision: D38722496

fbshipit-source-id: 5f9effeeb5a8cd62ea4a8edcde78f1e04768a9cf