[env_builder] Make `this` inside object methods read `IllegalThis`
Summary:
Changelog: [internal]

We already banned the references of this inside objects. Since this inside object methods are considered illegal, they should have any type. Since we already know exactly what type we will give to the reads of those thises, there is no point to create write entries for them.

In this diff, I created a new kind of value called `IllegalThis` that will be bind to the environment before visiting object methods, so that reads of this in object methods will read the `IllegalThis` value.

Reviewed By: mvitousek

Differential Revision: D38761274

fbshipit-source-id: 690a14c048ddae602c30f7058016b71ec2d55814