[refactor] Expose `infer_ast` and `infer_lib_file` that automatically chooses the environment based on the context
Summary:
Changelog: [internal]

It is very error prone to make every use site of Type_inference_js to check which environment we should use. In this diff, I changed the interface of `type_inference.ml` so that it will not expose `EnvInference` and `NewEnvInference`, but instead expose `infer_ast` and `infer_lib_file` that automatically chooses the right env based on the context.

Reviewed By: panagosg7

Differential Revision: D38800517

fbshipit-source-id: 3a0ca42edda3d17242698529bbd8df170f75f6c1