[new-env] Model `internal_name "exports"` in new-env
Summary:
Changelog: [internal]

In this diff, I will model in the new-env the last known internal variable that depends on the old-env: `internal_name "exports"`.

The operations performed by the old-env on this variable is fairly simple but very scope sensitive: `get_internal_var cx "exports" loc` and `set_internal_var cx "exports" loc`. The location passed into two functions is basically ignored, and the old-env uses the scope tracked by itself to decide with entry to read and write from.

I modeled this in the new env as a write entry with kind DeclareModuleExports and with the location of the entire declare module as the write location.

Reading of this entry only happens at the end of checking a declare module, where we can directly thread the location into the reader function.

Writing is more complicated. One write happens in `declare module.exports: some_type`, which can happen anywhere inside the declare module. This is the case where a scope aware solution is very desirable. Using the fact that reordered-env is the only environment we need to support from now on, I implemented the scope tracking inside env_resolution: when resolving a declare module, the current declare_module_export_loc is added to the loc_env, and the new-env can use this location to figure out which entry to write to in a `set_module_exports` call.

Reviewed By: panagosg7

Differential Revision: D38772143

fbshipit-source-id: 99bca72890e394560dad77cf9d94146096d8af4b